+++ CHANGELOG.md
@@ -6,6 +6,7 @@ QuiltPlus fixes
 
 - Namespace.pkg_name() returns the package name
 - Load Manifests lazily
+- Allow getting not-yet-existing Namespaces from a Registry
 
 ## 0.3.0 (2023-07-27)